+++ app/Widgets/DonorsWidget.php
@@ -3,7 +3,9 @@
 namespace App\Widgets;
 
 use App\Donor;
+use App\Donation;
 use Illuminate\Support\Facades\Auth;
+use Carbon\Carbon;
 
 class DonorsWidget implements Widget
 {
@@ -20,7 +22,9 @@ class DonorsWidget implements Widget
     function args(): array {
         return [
             'num_donors' => Donor::count(),
-			'latest_donor' => Donor::orderBy('created_at', 'desc')->first(),
+            'num_donations_month' => Donation::whereDate('date', '>', Carbon::now()->startOfMonth())->count(),
+            'num_donations_year' => Donation::whereDate('date', '>', Carbon::now()->startOfYear())->count(),
+            'num_donations_total' => Donation::count(),
         ];
     }
 }