+++ resources/views/dashboard/widgets/donors.blade.php
@@ -6,9 +6,11 @@
     <div class="card-body pb-2">
         <p>
             @lang('donations.donors_in_db', [ 'num_donors' => $num_donors ])<br>
-            @isset($latest_donor)
-                @lang('donations.newest_donor_is', [ 'link' => route('donors.show', $latest_donor), 'name' => $latest_donor->name ])
-            @endisset
+            @lang('donations.donations_in_db', [ 
+                'num_month' => $num_donations_month,
+                'num_year' => $num_donations_year,
+                'num_total' => $num_donations_total
+            ])<br>
         </p>
     </div>
 </div>